refactor AnimatedProgressbar -> AnimatedProgressProvider
